pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ninjamcp-python"
version = "1.3.0"
description = "MCP Server for NinjaRMM API v2 - Python Implementation"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["mcp", "ninjarmm", "api", "rmm"]
authors = [
    {name = "Chris Goetz", email = "goetzcj@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/goetzcj/ninjamcp-python"
Repository = "https://github.com/goetzcj/ninjamcp-python.git"
Issues = "https://github.com/goetzcj/ninjamcp-python/issues"

[project.scripts]
ninjamcp-python = "ninjarmm_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
strict_optional = true

[tool.isort]
profile = "black"
line_length = 100
